Rename id modules.
